+++ src/main/kotlin/com/codersgate/ticketraider/global/infra/security/PasswordEncoderConfig.kt
@@ -1,4 +1,4 @@
-package com.codersgate.ticketraider.infra.security
+package com.codersgate.ticketraider.global.infra.security
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
